fix : devsite-1574 -> codeblock issues
