+++ hlx_statics/blocks/codeblock/codeblock.css
@@ -27,6 +27,10 @@ main div.codeblock-wrapper div.codeblock .control-bar {
   justify-content: space-between;
 }
 
+main div .tabs-list {
+  margin-left: 30px;
+}
+
 /* copied from https://raw.githubusercontent.com/adobe/aem-block-collection/main/blocks/tabs/tabs.css, then edited */
 
 main div.codeblock-wrapper div.codeblock .control-bar,
@@ -35,7 +39,6 @@ main div.codeblock-wrapper div.codeblock .control-bar .tabs-list {
   gap: 16px;
   max-width: 100%;
   overflow-x: auto;
-  margin: 0 16px;
 }
 
 main div.codeblock-wrapper div.codeblock .control-bar select,
@@ -55,7 +58,6 @@ main div.codeblock-wrapper div.codeblock .control-bar .tabs-list button {
   cursor: pointer;
   background-color: initial;
   border-bottom: 3px solid transparent;
-  padding: 16px 0;
 }
 
 main div.codeblock-wrapper div.codeblock .control-bar select:hover,
