+++ hlx_statics/styles/prism.css
@@ -5,7 +5,7 @@ pre[class*=language-] {
     color: rgb(235, 235, 235);
     background: 0 0;
     font-family: Consolas, Monaco, 'Andale Mono', 'Ubuntu Mono', monospace;
-    font-size: 1em;
+    font-size: 14px;
     text-align: left;
     white-space: pre;
     word-spacing: normal;
@@ -25,7 +25,7 @@ pre[class*=language-] {
     padding: 1em;
     margin: 0;
     overflow: auto;
-    background: rgb(15, 15, 15);
+    background: #323232;
 }
 
 .token.comment,
